+++ packages/remix-server-runtime/data.ts
@@ -1,17 +1,11 @@
-import {
-  redirect,
-  json,
-  isDeferredData,
-  isResponse,
-  isRedirectStatusCode,
-} from "./responses";
+import { redirect, isDeferredData, isRedirectStatusCode } from "./responses";
 import type {
   ActionFunction,
   ActionFunctionArgs,
   LoaderFunction,
   LoaderFunctionArgs,
-  ResponseStub,
 } from "./routeModules";
+import type { ResponseStub } from "./single-fetch";
 
 /**
  * An object of unknown type for route loaders and actions provided by the
@@ -34,7 +28,6 @@ export async function callRouteAction({
   params,
   request,
   routeId,
-  singleFetch,
   response,
 }: {
   request: Request;
@@ -42,8 +35,7 @@ export async function callRouteAction({
   params: ActionFunctionArgs["params"];
   loadContext: AppLoadContext;
   routeId: string;
-  singleFetch: boolean;
-  response?: ResponseStub;
+  response: ResponseStub;
 }) {
   let result = await action({
     request: stripDataParam(stripIndexParam(request)),
@@ -59,12 +51,7 @@ export async function callRouteAction({
     );
   }
 
-  // Allow naked object returns when single fetch is enabled
-  if (singleFetch) {
-    return result;
-  }
-
-  return isResponse(result) ? result : json(result);
+  return result;
 }
 
 export async function callRouteLoader({
@@ -73,7 +60,6 @@ export async function callRouteLoader({
   params,
   request,
   routeId,
-  singleFetch,
   response,
 }: {
   request: Request;
@@ -81,8 +67,7 @@ export async function callRouteLoader({
   params: LoaderFunctionArgs["params"];
   loadContext: AppLoadContext;
   routeId: string;
-  singleFetch: boolean;
-  response?: ResponseStub;
+  response: ResponseStub;
 }) {
   let result = await loader({
     request: stripDataParam(stripIndexParam(request)),
@@ -108,12 +93,7 @@ export async function callRouteLoader({
     return result;
   }
 
-  // Allow naked object returns when single fetch is enabled
-  if (singleFetch) {
-    return result;
-  }
-
-  return isResponse(result) ? result : json(result);
+  return result;
 }
 
 // TODO: Document these search params better